refactor: CustomerService와 Repository 분리
CustomerService에서 데이터 접근 로직을 분리하여 CustomerRepository를 생성하였습니다
