+++ src/main/java/me/kimihiqq/vouchermanagement/domain/customer/service/CustomerService.java
@@ -6,11 +6,8 @@ import java.util.List;
 
 public interface CustomerService {
 
-    void init();
-
     Customer checkCustomer(String customerId);
 
     List<Customer> getBlacklistedCustomers();
-
 
 }
